requirements.txt
# ================================
# CORE DEPENDENCIES ONLY
# ================================

# Data manipulation
pandas==2.1.4
numpy==1.26.2

# Parquet support
pyarrow==14.0.2

# Config files  
PyYAML==6.0.1

# Testing
pytest==7.4.4